move docker package to pkg/container/docker
