+++ pkg/container/docker/kill.go
